Adding option for scaling of multiple users.
WIP push of adding the functionality that allows the user to specify
their own names for each person contributing to this bill.

So far I'm mostly done with the option of adding payers to a grocery
list.
TODO:
Be able to remove payers
Clear the payer list on opening a new list
...